refactor: client more logical layout added pages folder
