+++ client/src/Routes.module.css
@@ -1,0 +1,5 @@
+.main_layout {
+  margin: 4rem auto;
+  max-width: 1200px;
+  width: 100%;
+}
