--- client/src/components/App.module.css
@@ -1,39 +1,0 @@
-body {
-  margin: 20px;
-}
-
-.content {
-  grid-area: content;
-}
-
-.header {
-  grid-area: header;
-}
-
-.footer {
-  grid-area: footer;
-}
-
-.wrapper {
-  max-width: 1200px;
-  width: 100%;
-  margin: 0 auto;
-}
-
-/* input,
-select,
-textarea {
-  padding: 0.4em 0.2em;
-  font-size: 1.2rem;
-}
-
-.active {
-  font-weight: bold;
-}
-
-.form {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-}  */
