--- client/src/components/Navbar.module.css
@@ -1,32 +1,0 @@
-nav {
-  margin: 0 auto;
-  width: 100%;
-  text-align: center;
-  align-items: right;
-  background: #000;
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-auto-rows: auto;
-  grid-gap: 1em;
-}
-
-.navLinks {
-  font-size: 1.2em;
-  text-transform: uppercase;
-  text-decoration: none;
-  color: #fff;
-  padding: 1em;
-}
-
-.navLinks:hover {
-  color: white;
-  border: 1px solid white;
-}
-
-@media screen and (max-width: 900px) {
-  nav {
-    grid-template-columns: 100%;
-    grid-template-rows: auto;
-    grid-gap: 1em;
-  }
-}
