+++ client/src/components/navigation/Navbar.module.css
@@ -1,0 +1,45 @@
+.pageheader {
+  position: fixed;
+  left: 0;
+  top: 0;
+  height: 3.5rem;
+  background: #000;
+  width: 100%;
+  padding: 0 2rem;
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  grid-template-areas: 'logo logo logo nav';
+  grid-auto-rows: auto;
+  grid-gap: 1rem;
+}
+
+.pagenav {
+  grid-area: nav;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.navLinks {
+  font-size: 1.2rem;
+  text-transform: uppercase;
+  color: #fff;
+  padding: 0.5rem;
+}
+
+.navLinks:hover,
+.navLinks:active,
+.navLinks a:hover,
+.navLinks a:active,
+.navLinks a.active {
+  color: green;
+  text-decoration: none;
+}
+
+@media screen and (max-width: 900px) {
+  nav {
+    grid-template-columns: 100%;
+    grid-template-rows: auto;
+    grid-gap: 1rem;
+  }
+}
