+++ client/src/pages/Landing.module.css
@@ -1,0 +1,5 @@
+h1,
+h2,
+h3 {
+  text-align: center;
+}
